requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
streamlit>=1.30.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0 
